+++ pkg/types/config.go
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-package consensus
+package types
 
 import (
 	"time"
